Revert "avoid unsafe cast to int"
This reverts commit 44da0ea183b87fd8514048ecd510ea57fd3f2b08.

+++ poi-ooxml/src/main/java/org/apache/poi/xssf/model/SharedStringsTable.java
@@ -124,8 +124,8 @@ public class SharedStringsTable extends POIXMLDocumentPart implements SharedStri
             int cnt = 0;
             _sstDoc = SstDocument.Factory.parse(is, DEFAULT_XML_OPTIONS);
             CTSst sst = _sstDoc.getSst();
-            count = Math.toIntExact(sst.getCount());
-            uniqueCount = Math.toIntExact(sst.getUniqueCount());
+            count = (int)sst.getCount();
+            uniqueCount = (int)sst.getUniqueCount();
             //noinspection deprecation
             for (CTRst st : sst.getSiArray()) {
                 stmap.put(xmlText(st), cnt);
